[#744] Refactor UTXO fetching code in CompactBlockProcessor
Closes #744

- Added `UTXOFetcher` protocol and `UTXOFetcherImpl` struct.
